Join tests in same file
